+++ lambdas/src/service/Service.ts
@@ -1,8 +1,11 @@
+import { HealthStatus } from '../utils/ControllerUtils'
+
 /**
  * This version of the service can tell clients about the status of the base Lambdas Service.
  */
 export interface LambdasService {
   getStatus(): Promise<ServerStatus>
+  getHealth(): Promise<Record<string, HealthStatus>>
 }
 
 export type Timestamp = number
